+++ src/Users/Policies/UserPolicy.php
@@ -1,0 +1,64 @@
+<?php
+
+namespace Enraiged\Users\Policies;
+
+use Enraiged\Users\Models\User;
+// use Illuminate\Auth\Access\HandlesAuthorization;
+
+class UserPolicy
+{
+    // use HandlesAuthorization;
+
+    /**
+     *  @param  \Enraiged\Users\Models\User  $auth
+     *  @param  \Enraiged\Users\Models\User  $user
+     *  @return bool
+     */
+    public function delete(User $auth, User $user)
+    {
+        if (!is_null($user->deleted_at) || $user->isProtected || ($user->isMyself && !$user->allowSelfDelete)) {
+            return false;
+        }
+
+        return $user->isMyself && $user->allowSelfDelete;
+    }
+
+    /**
+     *  @param  \Enraiged\Users\Models\User  $auth
+     *  @param  \Enraiged\Users\Models\User  $user
+     *  @return bool
+     */
+    public function edit(User $auth, User $user)
+    {
+        return $user->isMyself;
+    }
+
+    /**
+     *  @param  \Enraiged\Users\Models\User  $auth
+     *  @return bool
+     */
+    public function index(User $auth)
+    {
+        return $auth->exists;
+    }
+
+    /**
+     *  @param  \Enraiged\Users\Models\User  $auth
+     *  @param  \Enraiged\Users\Models\User  $user
+     *  @return bool
+     */
+    public function show(User $auth, User $user)
+    {
+        return $user->isMyself;
+    }
+
+    /**
+     *  @param  \Enraiged\Users\Models\User  $auth
+     *  @param  \Enraiged\Users\Models\User  $user
+     *  @return bool
+     */
+    public function update(User $auth, User $user)
+    {
+        return $user->isMyself;
+    }
+}